New: Connect to Trakt first + more checks for outdated authorization
For the authorization it is now checked from a previous json file if
the acces codes are still valid. If not a refresh is tried.

The trakt authentication is moved at the start of the script, so it will
break directly if authentication was not successful and not after
grabbing all tv show data.

Also relates to #9